refactor(NavLink): rename NavItem to NavLink
