+++ src/components/Nav/NavLink.js
@@ -1,14 +1,16 @@
+/* eslint-disable jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events -- links are interactable, but eslint cannot statically check it here. */
 import Link from 'next/link'
 
 
 
 
-export default function NavItem (props) {
+export default function NavLink (props) {
   const {
     children,
     className,
     disabled,
     external,
+    onClick,
     ...restProps
   } = props
 
@@ -17,7 +19,7 @@ export default function NavItem (props) {
       {
         external
           ? (
-            <a {...restProps} className={[className, { disabled }]}>
+            <a {...restProps} className={[className, { disabled }]} onClick={onClick}>
               <span>
                 {children}
               </span>
@@ -25,7 +27,7 @@ export default function NavItem (props) {
           )
           : (
             <Link {...restProps}>
-              <a className={[className, { disabled }]}>
+              <a className={[className, { disabled }]} onClick={onClick}>
                 <span>
                   {children}
                 </span>
