+++ src/components/Nav/index.js
@@ -1,4 +1,4 @@
 export { default as Nav } from './Nav'
-export { default as NavItem } from './NavItem'
-export { default as ExternalNavItem } from './ExternalNavItem'
+export { default as NavLink } from './NavLink'
+export { default as NavSection } from './NavSection'
 export { default as SubNav } from './SubNav'
